add: match_song_to_playlist, get_titles, get_songs_in_playlist
Again, some methods should fail loudly to know when to log and notify the user that their request failed.
Likewise, add_song_to_playlist notifies the user when their request to add a song also created the playlist to add the song to.
This means we currently prefer to allow the user to add songs to new playlists easily (by creating them just before adding),
which also means a new playlist is created when the user misspells the name.