change: dosctrings update
For most cogs, update to a new format. Slightly improves readability, simplifies the format (specially when an argument uses more than one line)
and renders better on vscode's hints.
For SongQueue, update to a sphinx-compatible format to auto-generate the docs with it.
Since this class doesn't use decorators that mess up with __doc__, and it is a useful API for anyone wanting to modify the bot,
a good documentation can be easily generated with this change.