change to protect global variables from imports, Meme now stores its sessions
Following the trend to protect variables when imported, initial setup is moved to setup function.
No longer create and close a Reddit instance each time a meme is requested. I hope this doesn't impact performance when connected for a long time,
but it would seem it speeds up sending memes after the first one.