add new version of docs
The docs are now written in reStructuredText in order to use sphinx and read the docs to build and host the documentation.
A bit of order is added to .env.example and to the table of package dependencies.
The index (main page) which was created from the README is decluttered, with the introduction and 'instruction on how to create the bot' parts moved to the Getting started pages.