change docs theme to furo
Change the theme used by sphinx to furo.
Add a couple of extensions to add a copy button to code blocks and have inline_tabs, to present two versions of a command.
The old documentation is moved to another folder, but remains intact for now.
There shouldn't be any need for those docs in the future, but there's also no need to delete them yet.
Add the locale settings to conf, anticipating the translation of the docs.